Move master recipe list to app component
The app component now binds the recipe list to the recipe list
component. I think it's easier to re-render the recipe list this way
when new recipes are added, etc.

Since it's the app component that holds the recipe list data, it now
calls `getRecipeList()` directly on init. The initialization is now a
hidden implementation detail (and it just feels better designed this
way).