Enable adding recipes via the view
There's now a form with which the user can add (or possible edit)
recipes. A link in the recipe list displays the form.

I added an observable in the recipe service so the app component can
know when the recipe list is updated by any part of the app (I thought I
could @Output new recipes from the dialog to the app root, but it turns
out I can't do that because the dialog is in the router outlet).

Now I'm thinking if I should make a separate service for updating the
local storage, and have it subscribe to recipe updates as well.

I'm also thinking of moving back the recipe list from the app component
down to the recipe list component, since my initial assumption is wrong.