Allow recipe edits
I reused the RecipeDialog component for making recipe edits, but now I
had to check if the router link has a recipe id in it. If so, it'll edit
the recipe with that id. Otherwise it's the same old Add Recipe
function.

I don't know, the checks that I made looks clumsy and "not
Angular-like".