Layout recipe list
The recipe list now grows downward until it can't fit everything on the
screen, in which case it becomes a scrolling list. For that to work I
had to give all of its parent elements 100% height.

I also switched from <ul> to a <div> with a bunch of <a> descendants,
because I wanted the entire item on the screen to be clickable, not just
the text. I think it made my code less semantic. I could have achieved
the same effect with <ul> by changing the <li> paddings into <a> margins
(or paddings). But for now this is less of a hassle.

I think it's also better to give the elements classes rather than use
element selectors and rely on Angular's CSS encapsulation.